requirements.txt
amqp==5.2.0
APScheduler==3.10.4
asarPy==1.0.1
asgiref==3.8.1
aspose-zip==24.6.0
Babel==2.15.0
bcrypt==4.2.0
billiard==4.2.0
blinker==1.8.2
cachelib==0.13.0
celery==5.4.0
click==8.1.7
click-didyoumean==0.3.1
click-plugins==1.1.1
click-repl==0.3.0
dnspython==2.6.1
email_validator==2.2.0
filelock==3.15.4
Flask==2.3.3
Flask-BabelEx==0.9.4
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Mail==0.10.0
Flask-Principal==0.4.0
Flask-Security==3.0.0
Flask-Session==0.8.0
Flask-SQLAlchemy==3.1.1
flask-talisman==1.1.0
Flask-WTF==1.2.1
fsspec==2024.6.1
greenlet==3.0.3
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
joblib==1.4.2
kombu==5.3.7
llvmlite==0.43.0
MarkupSafe==2.1.5
mpmath==1.3.0
msgspec==0.18.6
networkx==3.3
numba
numpy
packaging==24.1
pandas
passlib==1.7.4
pillow==10.4.0
plotly==5.23.0
prompt_toolkit==3.0.47
pynndescent==0.5.13
python-dateutil==2.9.0.post0
pytz==2024.1
redis==5.0.7
scikit-learn
scipy
six==1.16.0
speaklater==1.3
SQLAlchemy==2.0.31
sympy==1.13.1
tenacity==9.0.0
threadpoolctl==3.5.0
torch==2.4.0
torchvision==0.19.0
tqdm==4.66.4
umap-learn
vine==5.1.0
wcwidth==0.2.13
Werkzeug==3.0.3
WTForms==3.1.2
flask[async]
flask_security   